+++ src/line/condition.rs
@@ -31,7 +31,7 @@
 //! more information.
 
 use crate::{
-    error::parse::address::InvalidAddressError,
+    error::{parse::address::InvalidAddressError, utils::MetaData},
     knot::{Address, ValidateAddressData, ValidateAddresses},
     line::Variable,
 };
@@ -224,17 +224,20 @@ impl ConditionBuilder {
 impl ValidateAddresses for Condition {
     fn validate(
         &mut self,
+        errors: &mut Vec<InvalidAddressError>,
+        meta_data: &MetaData,
         current_address: &Address,
         data: &ValidateAddressData,
-    ) -> Result<(), InvalidAddressError> {
-        self.root.kind.validate(current_address, data)?;
+    ) {
+        self.root
+            .kind
+            .validate(errors, meta_data, current_address, data);
 
-        self.items
-            .iter_mut()
-            .map(|item| match item {
-                AndOr::And(item) | AndOr::Or(item) => item.kind.validate(current_address, data),
-            })
-            .collect()
+        self.items.iter_mut().for_each(|item| match item {
+            AndOr::And(item) | AndOr::Or(item) => {
+                item.kind.validate(errors, meta_data, current_address, data)
+            }
+        });
     }
 
     #[cfg(test)]
@@ -249,13 +252,17 @@ impl ValidateAddresses for Condition {
 impl ValidateAddresses for ConditionKind {
     fn validate(
         &mut self,
+        errors: &mut Vec<InvalidAddressError>,
+        meta_data: &MetaData,
         current_address: &Address,
         data: &ValidateAddressData,
-    ) -> Result<(), InvalidAddressError> {
+    ) {
         match self {
-            ConditionKind::True | ConditionKind::False => Ok(()),
-            ConditionKind::Nested(condition) => condition.validate(current_address, data),
-            ConditionKind::Single(kind) => kind.validate(current_address, data),
+            ConditionKind::True | ConditionKind::False => (),
+            ConditionKind::Nested(condition) => {
+                condition.validate(errors, meta_data, current_address, data)
+            }
+            ConditionKind::Single(kind) => kind.validate(errors, meta_data, current_address, data),
         }
     }
 
@@ -272,18 +279,23 @@ impl ValidateAddresses for ConditionKind {
 impl ValidateAddresses for StoryCondition {
     fn validate(
         &mut self,
+        errors: &mut Vec<InvalidAddressError>,
+        meta_data: &MetaData,
         current_address: &Address,
         data: &ValidateAddressData,
-    ) -> Result<(), InvalidAddressError> {
+    ) {
         match self {
             StoryCondition::Comparison {
                 ref mut lhs_variable,
                 ref mut rhs_variable,
                 ..
-            } => lhs_variable
-                .validate(current_address, data)
-                .and_then(|_| rhs_variable.validate(current_address, data)),
-            StoryCondition::IsTrueLike { variable } => variable.validate(current_address, data),
+            } => {
+                lhs_variable.validate(errors, meta_data, current_address, data);
+                rhs_variable.validate(errors, meta_data, current_address, data);
+            }
+            StoryCondition::IsTrueLike { variable } => {
+                variable.validate(errors, meta_data, current_address, data)
+            }
         }
     }